Added launch mode to the script for login screen. Uses the default background and draws the clock. Just run the script with ANY command line arguments. Example: ./lock.sh a

+++ .scripts/lock.sh
@@ -1,11 +1,7 @@
-#!/bin/bash
+#!/usr/bin/env bash
 # Author: Rajex 
 # Dependencies: ImageMagick, i3lock-color-git, sed, grep
 set -o errexit -o noclobber -o nounset
-
-effect=(-scale 10% -scale 1000%)
-image=$(mktemp).png
-shot=(import -window root)
 
 letterEnteredColor=d23c3dff
 letterRemovedColor=d23c3dff
@@ -13,8 +9,9 @@ passwordCorrect=00000000
 passwordIncorrect=d23c3dff
 background=00000000
 foreground=ffffffff
-i3lock_cmd=(i3lock -i "$image" -t -n \
-    --timepos="x-90:h-ch+30" \
+image="$HOME/wallpaper.png"
+i3lock_cmd=(i3lock -i "$image" -t -n)
+params=(--timepos="x-90:h-ch+30" \
     --datepos="tx+24:ty+25" \
     --clock --datestr "Type password to unlock..." \
     --insidecolor=$background --ringcolor=$foreground --line-uses-inside \
@@ -25,9 +22,19 @@ i3lock_cmd=(i3lock -i "$image" -t -n \
     --textcolor="$foreground" --timecolor="$foreground" --datecolor="$foreground" \
     --force-clock)
 
-trap 'rm -f "$image"' EXIT
-command -- "${shot[@]}" "$image"
+if [[ $# -eq 1 ]] ; then
+    effect=()
+    "${i3lock_cmd[@]}" "${params[@]}"
+    exit
+fi
 
+effect=(-scale 10% -scale 1000%)
+image=$(mktemp).png
+shot=(import -window root)
+i3lock_cmd=(i3lock -i "$image" -t -n)
+
+command -- "${shot[@]}" "$image"
+trap 'rm -f "$image"' EXIT
 dom=$(convert $image -scale 1x1\! -modulate 100,200,120 -format '%[pixel:u]' info:- | sed 's/[^0-9,]*//g')
 transparency=0.85
 
@@ -38,8 +45,8 @@ for RES in $SR; do
     CX=$((${SRA[2]} + 25))                                                  
     CY=$((${SRA[1]} - 30))                                                  
     rectangles+="rectangle $CX,$CY $((CX+300)),$((CY-80)) "                 
-done        
+done
 
 convert "$image" "${effect[@]}" -draw "fill rgba($dom, $transparency) $rectangles" "$image"
 
-"${i3lock_cmd[@]}"
+"${i3lock_cmd[@]}" "${params[@]}"
